scientific-advisor-core/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.1
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
